src/services/strategy_engine.py
"""
Service for securely processing and applying strategy definitions.
"""
import pandas as pd
from typing import Dict, Any, Type
from backtesting import Strategy

__all__ = ["StrategyEngine"]


class StrategyEngine:
    """
    Translates a JSON strategy definition into an executable backtesting.py Strategy.
    This component is critical for security as it avoids using eval().
    """
    def __init__(self) -> None:
        pass

    def create_strategy_class(self, strategy_def: Dict[str, Any]) -> Type[Strategy]:
        """
        Dynamically creates a backtesting.py Strategy class from a definition.

        Returns:
            A new class that inherits from backtesting.py's Strategy.
        """
        # This is a placeholder implementation.
        class _DynamicStrategy(Strategy):  # type: ignore[misc]
            def init(self) -> None:
                pass

            def next(self) -> None:
                pass

        return _DynamicStrategy
